requirements.txt
python-telegram-bot
facebook-sdk
pyyaml